tutorial-1: improve dependency versionning recommendation
Real-life shows that using the x.y.z form causes interoperability
issues and gets in the way of using the latest bindgen version.

NOTE: there is some interaction here with the semver semantics as
interpreted by Cargo: to fully work it requires a commitment to to
introduce (too many) breaking changes between 0.x versions.

NOTE: the root cause of the problem seems to be building all
dependency crates within the app's workspace: outdated build-deps for
a given lib is a problem, but should not prevent the build of the app.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>